+++ SDL_Template/include/DColors.h
@@ -13,6 +13,6 @@
 
 template <typename T>
 SDL_Color BlendColor(const SDL_Color& c0, const SDL_Color& c1, T percentage) {
-	SDL_Color output = { (c1.r - c0.r) * percentage + c0.r, (c1.g - c0.g) * percentage + c0.g, (c1.b - c0.b) * percentage + c0.b, (c1.a - c0.a) * percentage + c0.a };
+	SDL_Color output = { Uint8((c1.r - c0.r) * percentage + c0.r), Uint8((c1.g - c0.g) * percentage + c0.g), Uint8((c1.b - c0.b) * percentage + c0.b), Uint8((c1.a - c0.a) * percentage + c0.a) };
 	return output;
 }
